refactor(android-ui): move search bar to shared components
