+++ androidApp/src/main/java/com/abelvolpi/kmptravelapp/android/presentation/ui/explore/ExploreScreen.kt
@@ -10,8 +10,6 @@ import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
@@ -20,26 +18,13 @@ import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.SearchBar
-import androidx.compose.material3.SearchBarDefaults
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
@@ -50,10 +35,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
 import coil.decode.SvgDecoder
 import coil.request.ImageRequest
-import com.abelvolpi.kmptravelapp.android.R
 import com.abelvolpi.kmptravelapp.android.presentation.components.LoadingIndicator
+import com.abelvolpi.kmptravelapp.android.presentation.components.SearchBarComponent
 import com.abelvolpi.kmptravelapp.android.presentation.theme.MyApplicationTheme
-import com.abelvolpi.kmptravelapp.android.presentation.theme.secondaryColor
 import com.abelvolpi.kmptravelapp.android.presentation.theme.tertiaryColor
 import com.abelvolpi.kmptravelapp.android.presentation.ui.home.HomeScreen
 import com.abelvolpi.kmptravelapp.data.model.Category
@@ -102,64 +86,6 @@ fun ExploreUI(
             places = exploreUIData.places,
             onPlaceClicked = onPlaceClicked
         )
-    }
-}
-
-
-
-//TODO move to shared components
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun SearchBarComponent() {
-    var text by remember { mutableStateOf("") }
-    var active by remember { mutableStateOf(false) }
-    SearchBar(
-        modifier = Modifier
-            .fillMaxWidth()
-            .heightIn(0.dp, LocalConfiguration.current.screenHeightDp.dp)
-            .padding(horizontal = 20.dp),
-        shape = RoundedCornerShape(15.dp, 15.dp, 15.dp, 15.dp),
-        colors = SearchBarDefaults.colors(
-            containerColor = secondaryColor,
-            inputFieldColors = TextFieldDefaults.colors(
-                focusedTextColor = Color.White,
-                unfocusedTextColor = Color.White,
-                unfocusedPlaceholderColor = Color.White,
-                focusedPlaceholderColor = Color.White,
-            )
-        ),
-        query = text,
-        onQueryChange = { text = it },
-        onSearch = { active = false },
-        active = active,
-        onActiveChange = { active = it },
-        placeholder = {
-            Text(text = "Search")
-        },
-        leadingIcon = {
-            Icon(
-                painter = painterResource(id = R.drawable.search_icon),
-                contentDescription = "Pesquisar",
-                tint = Color.White
-            )
-        },
-        trailingIcon = {
-            if (active) {
-                Icon(
-                    modifier = Modifier.clickable {
-                        if (text.isNotEmpty()) {
-                            text = ""
-                        } else {
-                            active = false
-                        }
-                    },
-                    imageVector = Icons.Default.Close,
-                    contentDescription = "Clear",
-                    tint = Color.White
-                )
-            }
-        }
-    ) {
     }
 }
